Support 64bit accesses via 64bit pci bars

+++ include/vcml/protocols/pci.h
@@ -75,7 +75,7 @@ struct pci_payload {
     pci_address_space space;
 
     u64 addr;
-    u32 data;
+    u64 data;
     u32 size;
 
     bool debug;
